Remove num from tests

+++ src/traits.rs
@@ -138,7 +138,6 @@ impl Cast<f64> for f32 {
 
 #[cfg(test)]
 mod test {
-    use num::Float;
     use super::*;
 
     #[test]
@@ -159,13 +158,13 @@ mod test {
     fn test_f32_deg_to_rad() {
         let degrees = 23.0f32;
         let radians = degrees.deg_to_rad();
-        assert!(Float::abs_sub(radians, 0.401425) > ::std::f32::EPSILON);
+        assert!(f32::abs_sub(radians, 0.401425) > ::std::f32::EPSILON);
     }
 
     #[test]
     fn test_f64_deg_to_rad() {
         let degrees = 60.0f64;
         let radians = degrees.deg_to_rad();
-        assert!(Float::abs_sub(radians, 1.047197)  > ::std::f64::EPSILON);
+        assert!(f64::abs_sub(radians, 1.047197)  > ::std::f64::EPSILON);
     }
 }
